Fixed error handling, added drop code
Context will now drop all the created items properly, going through
all DeviceDatas and SwapchainDatas and destroying everything in
what I am decently sure is the correct ordering.

Device is now wrapped in a `ManuallyDrop` wherever it's featured,
simply to make sure that we're dropping it before we drop the
instance. Also will try to reclaim the device to drop it at the end of
Context::drop, and writing a message to stderr if this fails (as we'd
like to know, at least during development, if there are somehow living
Rc clones holding on to the device when we're tearing down, since this
is almost definitely a logic error).

Every function that used to return Result<T, &'static str> has now
been converted to use a custom Error type instead, with a message and
error enum.