A lot of changes
Fixed error handling to be more in line with modern rust
standards (i.e. one error enum that implements display)

Added the resources/ module/directory, for the low level resource
management (memory, geometry buffers, staging buffers, textures, etc.)

Added nalgebra and take_mut to dependencies.

Added a `Tracker` enum that changes state whenever a mutable reference
to its underlying data is acquired, this is going to be used for
making sure what needs to be reuploaded to the GPU for draw calls

Added a SceneTree struct that will hold all of our geometry and
transforms.

Added src/geometry.rs (not to be confused with
src/graphics/resources/geometry.rs) for our geometry/math structs,
i.e. the matrices, vectors, quads, etc.

Right now the code produces a lot of warnings, the idea behind not
removing them is to serve as a reminder of where things need
fixes/improvements, as it would take way too long to always make sure
everything is pristine for each commit.